+++ apps/web/src/app/AppWrapper/index.tsx
@@ -6,7 +6,7 @@ import {
   ConfigProvider,
   usePlatform
 } from '@vkontakte/vkui'
-import { router } from '../../routes'
+import { router } from '../../routes/router'
 import AuthProvider from './AuthProvider.tsx'
 
 import App from './App'
