+++ apps/web/src/routes/constants.ts
@@ -1,0 +1,16 @@
+export const VIEW_SCHEDULE = 'schedule'
+export const VIEW_CONTACTS = 'contacts'
+export const VIEW_MARKS = 'marks'
+export const VIEW_SETTINGS = 'settings'
+export const VIEW_ATTESTATION = 'attestation'
+export const VIEW_NOTIFICATIONS = 'notifications'
+
+export const MAIN_SETTINGS = 'login'
+
+export const PAGE_MAIN = '/'
+export const PAGE_SCHEDULE = `/${VIEW_SCHEDULE}`
+export const PAGE_CONTACTS = `/${VIEW_CONTACTS}`
+export const PAGE_MARKS = `/${VIEW_MARKS}`
+export const PAGE_SETTINGS = `/${VIEW_SETTINGS}`
+export const PAGE_ATTESTATION = `/${VIEW_ATTESTATION}`
+export const PAGE_NOTIFICATIONS = `/${VIEW_NOTIFICATIONS}`
