+++ apps/web/src/routes/router.ts
@@ -1,0 +1,4 @@
+import { createHashRouter } from '@vkontakte/vk-mini-apps-router'
+import { routes } from './routes.ts'
+
+export const router = createHashRouter(routes)
